+++ src/main/scala/clustered_common/Body.scala
@@ -5,14 +5,14 @@ import math.Vec2
 
 class Body(
             id: String,
-            mass: BigDecimal,
+            mass: Double,
             startPosition: Vec2,
             startVelocity: Vec2
           ) extends AbstractBody(id, mass, startPosition, startVelocity) {
 
   def changePosition(changeVec: Vec2): Unit = this.position = this.position + changeVec
 
-  def findNewCluster(currDist: BigDecimal, neighbours: Set[Object]): Option[Object] = {
+  def findNewCluster(currDist: Double, neighbours: Set[Object]): Option[Object] = {
     val neighbourDescriptors = neighbours.map(neighbour => (neighbour, neighbour.position.distance(this.position)))
       .filter(neighbourDescriptor => neighbourDescriptor._2 < currDist)
       .toList
