+++ src/main/scala/divided/ClusterDescriptor.scala
@@ -4,6 +4,6 @@ import `object`.Object
 import math.Vec2
 
 case class ClusterDescriptor(id: String,
-                             mass: BigDecimal,
+                             mass: Double,
                              var position: Vec2,
                              var timestamp: Int) extends Object
