+++ src/main/scala/message/BodyDataUpdate.scala
@@ -2,4 +2,4 @@ package message
 
 import math.Vec2
 
-case class BodyDataUpdate(id: String, mass: BigDecimal, position: Vec2)
+case class BodyDataUpdate(id: String, mass: Double, position: Vec2)
