+++ src/main/scala/message/ClusteredDataInit.scala
@@ -2,4 +2,4 @@ package message
 
 import math.Vec2
 
-case class ClusteredDataInit (id: String, mass: BigDecimal, position: Vec2)
+case class ClusteredDataInit (id: String, mass: Double, position: Vec2)
