+++ src/main/scala/object/Object.scala
@@ -4,7 +4,7 @@ import math.Vec2
 
 trait Object {
   val id: String
-  val mass: BigDecimal
+  val mass: Double
   var position: Vec2
 
   override def equals(obj: Any): Boolean = {
